:sparkles: feat: implements home viewmodel

+++ app/src/main/java/com/riquelme/financecontrolapp/model/Order.kt
@@ -4,5 +4,5 @@ data class Order(
     val name: String,
     val price: Double,
     val date: String,
-    val entryOrExit: Long) {
+    val entryOrExit: Int) {
 }
